Pull channel into it's own module.
